Move DisplayablePlan to `datafusion_common`
