+++ datafusion/common/src/lib.rs
@@ -20,6 +20,7 @@ mod column;
 pub mod config;
 pub mod delta;
 mod dfschema;
+pub mod display;
 mod error;
 pub mod from_slice;
 pub mod parsers;
